cisco-mds-release-agent/requirements.txt
requests
beautifulsoup4
PyYAML
lxml
langchain
langsmith